requirements.txt
Flask
flask_cors
mysql-connector-python
bcrypt
sshtunnel
requests
python-dotenv
